+++ test/test_utils.dart
@@ -1,5 +1,5 @@
 /// Formats a `String` to better represent its underlying value.
-String formatInput(String input) {
+String formatInput(String? input) {
   if (input == null) {
     return '$input';
   } else if (input.isEmpty) {